Solved Bugs: 1. Auto route distance field on view hole page can have floating point value 2. Validation added on these fields 3. Machinery can have null values on capacity

+++ src/main/java/com/nicolrom/entities/Hole.java
@@ -57,7 +57,7 @@ public class Hole implements Serializable {
     private Area area;
 
     @Column
-    private Integer autoRouteDistance;
+    private Double autoRouteDistance;
 
     @Column
     private Integer autoStationaryTime;
@@ -172,11 +172,11 @@ public class Hole implements Serializable {
         this.area = area;
     }
 
-    public Integer getAutoRouteDistance() {
+    public Double getAutoRouteDistance() {
         return autoRouteDistance;
     }
 
-    public void setAutoRouteDistance(Integer autoRouteDistance) {
+    public void setAutoRouteDistance(Double autoRouteDistance) {
         this.autoRouteDistance = autoRouteDistance;
     }
